+++ src/backend/profileFiltersHelper.sanitaire.test.ts
@@ -134,6 +134,9 @@ function getBudgetFinanceSanitaire(): EntitéJuridiqueBudgetFinance {
     resultatNetComptable: 30,
     ratioDependanceFinanciere: 31,
     tauxDeCafNetSan: 32,
+    fondsDeRoulement: 33,
+    besoinFondsDeRoulement: 34,
+    tresorerie: 35,
   }
 }
 
@@ -241,6 +244,9 @@ function getBudgetFinanceProfile() {
     ratioDépendanceFinancière: "ok",
     tauxDeCafNette: "ok",
     allocationDeRessources: "ok",
+    fondsDeRoulement: "ok",
+    besoinFondsDeRoulement: "ok",
+    tresorerie: "ok",
   }
 }
 
@@ -857,6 +863,63 @@ describe("Filtre les informations de budget et finance des etablissement juridiq
     let etabSanitaireResult = getFullSanitaire();
     const profile = getFullProfile();
     profile.budgetEtFinance.tauxDeCafNette = "Ko";
+
+    // When
+    etabSanitaireResult = filterEtablissementSanitaire(etabSanitaireResult, profile);
+
+    // Then
+    expect(etabSanitaireResult.budgetFinance).toEqual([expectedBudgetAndFinance]);
+  })
+
+  it("retire les info sur les fonds de roulement si il n’y a pas les droits", () => {
+    // Given
+    const rawBudgetAndFinance = getBudgetFinanceSanitaire();
+    const expectedBudgetAndFinance = {
+      ...rawBudgetAndFinance,
+      fondsDeRoulement: "",
+    }
+
+    let etabSanitaireResult = getFullSanitaire();
+    const profile = getFullProfile();
+    profile.budgetEtFinance.fondsDeRoulement = "Ko";
+
+    // When
+    etabSanitaireResult = filterEtablissementSanitaire(etabSanitaireResult, profile);
+
+    // Then
+    expect(etabSanitaireResult.budgetFinance).toEqual([expectedBudgetAndFinance]);
+  })
+
+  it("retire les info sur le besoin en fonds de roulement si il n’y a pas les droits", () => {
+    // Given
+    const rawBudgetAndFinance = getBudgetFinanceSanitaire();
+    const expectedBudgetAndFinance = {
+      ...rawBudgetAndFinance,
+      besoinFondsDeRoulement: "",
+    }
+
+    let etabSanitaireResult = getFullSanitaire();
+    const profile = getFullProfile();
+    profile.budgetEtFinance.besoinFondsDeRoulement = "Ko";
+
+    // When
+    etabSanitaireResult = filterEtablissementSanitaire(etabSanitaireResult, profile);
+
+    // Then
+    expect(etabSanitaireResult.budgetFinance).toEqual([expectedBudgetAndFinance]);
+  })
+
+  it("retire les info sur la trésorerie si il n’y a pas les droits", () => {
+    // Given
+    const rawBudgetAndFinance = getBudgetFinanceSanitaire();
+    const expectedBudgetAndFinance = {
+      ...rawBudgetAndFinance,
+      tresorerie: "",
+    }
+
+    let etabSanitaireResult = getFullSanitaire();
+    const profile = getFullProfile();
+    profile.budgetEtFinance.tresorerie = "Ko";
 
     // When
     etabSanitaireResult = filterEtablissementSanitaire(etabSanitaireResult, profile);