+++ src/backend/profileFiltersHelper.ts
@@ -540,6 +540,9 @@ const filterBudgetFinanceEJ = (budgetFinance: any, profil: any) => {
     budget.resultatNetComptable = profil.résultatNetComptable === "ok" ? budget.resultatNetComptable : "";
     budget.ratioDependanceFinanciere = profil.ratioDépendanceFinancière === "ok" ? budget.ratioDependanceFinanciere : "";
     budget.tauxDeCafNetSan = profil.tauxDeCafNette === "ok" ? budget.tauxDeCafNetSan : "";
+    budget.fondsDeRoulement = profil.fondsDeRoulement === "ok" ? budget.fondsDeRoulement : "";
+    budget.besoinFondsDeRoulement = profil.besoinFondsDeRoulement === "ok" ? budget.besoinFondsDeRoulement : "";
+    budget.tresorerie = profil.tresorerie === "ok" ? budget.tresorerie : "";
   }
 
   return budgetFinance;