+++ CHANGELOG.md
@@ -20,11 +20,18 @@
 - vr-164: Placement des étiquettes de valeur sur les graphiques en barres verticales
 - vr-165: Modifications des couleurs des étiquettes de pourcentage dans le graphique répartition des effectifs
 
+- hel-996: Ajout des données concernant les fonds de roulement, les besoins en fonds de roulement et la trésorerie
+- hel-997: Ajout du graphique horizontal concernant les fonds de roulement
+- hel-998: Ajout du graphique horizontal concernant le besoin en fonds de roulement
+- hel-999: Ajout du graphique horizontal concernant la trésorerie
+- hel-1007: Ajout des nouvelles informations de budget dans la page de comparaison
+
 ## [1.8.3](https://github.com/DNUM-SocialGouv/helios/compare/release/1.8.2.1...release/1.8.3) (2026-03-09)
 
 ### Bug Fixes
 
 ### Features
+
 - vr-152: Renommage des colonnes de plusieurs tables
 
 ## [1.8.2.1](https://github.com/DNUM-SocialGouv/helios/compare/release/1.8.2...release/1.8.2.1) (2026-03-05)